kpm: add install command and fix TCP multi-connection
- cmd_kpm.c: implement kpm install <pkg> command
  - Reads /persist/kpm/packages_list.txt (name=path format)
  - Downloads package file from server via HTTP GET
  - Saves to /persist/kpm/packages/<name>
  - Updated usage message

- net.c / net.h: fix TCP reliability issues
  - net_tcp_recv: change maxlen from uint16_t to int (fix 65536 -> 0 overflow)
  - tcp_handle_ipv4_tcp: accumulate rx data instead of overwriting on each packet
  - net_http_get_to_buf: rewrite rx loop to drain all data before FIN closes it
  - net_tcp_connect: increment source port per connection (40000++)
    to avoid TIME_WAIT reuse rejection on consecutive requests

+++ include/kernel/drivers/net/net.h
@@ -14,7 +14,7 @@ void net_get_ipv4(uint32_t* ip_be, uint32_t* mask_be, uint32_t* gw_be);
 // --- TCP minimal API (tek bağlantı) ---
 int net_tcp_connect(uint32_t dst_ip_be, uint16_t dst_port);
 int net_tcp_send(const uint8_t* data, uint16_t len);
-int net_tcp_recv(uint8_t* out, uint16_t maxlen, uint32_t spin_timeout);
+int net_tcp_recv(uint8_t* out, int maxlen, uint32_t spin_timeout);
 int net_tcp_close(void);
 
 int net_http_get_to_buf(uint32_t ip_be, uint16_t port, const char* path,
